shiny/driver/mtldriver: switch bgra behavior by intel vs arm
