+++ shiny/driver/mtldriver/bgra.go
@@ -1,3 +1,6 @@
+//go:build arm64 && darwin
+// +build arm64,darwin
+
 package mtldriver
 
 import (
